+++ test/add_exclusion_contraint_test.rb
@@ -1,0 +1,16 @@
+require_relative "test_helper"
+
+class AddExclusionConstraintTest < Minitest::Test
+  def setup
+    skip unless ActiveRecord::VERSION::STRING.to_f >= 7.1
+    super
+  end
+
+  def test_basic
+    assert_unsafe AddExclusionConstraint
+  end
+
+  def test_new_table
+    assert_safe AddExclusionConstraintNewTable
+  end
+end