+++ test/test_helper.rb
@@ -40,6 +40,9 @@ schema_migration.create_table
 ActiveRecord::Schema.define do
   enable_extension "citext" if $adapter == "postgresql"
 
+  # for exclusion constraints
+  enable_extension "btree_gist" if $adapter == "postgresql"
+
   # for gen_random_uuid() in Postgres < 13
   enable_extension "pgcrypto" if $adapter == "postgresql"